+++ src/impls/null-or/null-or.d.ts
@@ -1,5 +1,0 @@
-import { SourceIterable, ResultIterable } from '../../types/iterable';
-
-declare function nullOr<T>(source: SourceIterable<T>): null | ResultIterable<T>;
-
-export { nullOr };